Fix shutdown crash when log/error callback is set
The rd_kafka_destroy method may call back into Haskell to perform some logging. This is not allowed when attaching the finalizer
using Foreign.ForeignPtr. Hence we switch to Foreign.Concurrent, which allows RdKafka to call back into Haskell. We also disable
closing the consumer automatically in the finalizer, this should be done by the application by calling `closeConsumer` explicitly.